cosmos_policy/experiments/robot/robotwin/requirements_robotwin_eval.txt
json-numpy>=2.0.0
numpy>=1.26.0
pillow>=11.0.0
requests>=2.32.0
